requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
qdrant-client
sentence-transformers
